project-board/src/main/java/com/project/projectb/service/ArticleService.java
package com.project.projectb.service;

import com.project.projectb.domain.type.SearchType;
import com.project.projectb.dto.ArticleDto;
import com.project.projectb.repository.ArticleRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.time.LocalDateTime;
import java.util.List;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleService {

    private final ArticleRepository articleRepository;


    @Transactional(readOnly = true)
    public List<ArticleDto> searchArticles(SearchType title, String searchKeyword) {
        return List.of(
        );
    }
}
